[M1] refactor: 답글 삭제 가능 여부 판단 역할을 답글 클래스로 이동
참고: https://github.com/next-step/java-bowling/pull/885

+++ src/main/java/qna/domain/AnswerWrapper.java
@@ -1,5 +1,7 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -15,12 +17,25 @@ public class AnswerWrapper {
         return answers;
     }
     
-    public List<DeleteHistory> deleteAll(){
+    public List<DeleteHistory> deleteAll(User loginUser) throws CannotDeleteException {
+        checkDeletable(loginUser);
         List<DeleteHistory> deleteHistories = new ArrayList<>();
         this.answers.forEach((answer)->{
             answer.setDeleted(true);
             deleteHistories.add(new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now()));
         });
         return deleteHistories;
+    }
+
+    private void checkDeletable(User loginUser) throws CannotDeleteException {
+        for (Answer answer : this.answers) {
+            checkAnswerOwner(loginUser, answer);
+        }
+    }
+
+    private void checkAnswerOwner(User loginUser, Answer answer) throws CannotDeleteException {
+        if (!answer.isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 }
